common components extracted
