+++ src/pages/communication-page/components/communication.components.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from 'ionic-angular';
-import { ComponentsModule } from '../../../../components/common/common-components.module';
 import { ProvidedEmailComponent } from './provided-email/provided-email';
 import { NewEmailComponent } from './new-email/new-email';
 import { PostalAddressComponent } from './postal-address/postal-address';
 import { PrivacyNoticeComponent } from './privacy-notice/privacy-notice';
 import { TranslateModule } from 'ng2-translate';
+import { ComponentsModule } from '../../../components/common/common-components.module';
 
 @NgModule({
   declarations: [
