+++ src/pages/communication-page/components/privacy-notice/privacy-notice.scss
