+++ src/pages/communication-page/components/privacy-notice/privacy-notice.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { TranslateService } from 'ng2-translate';
-import { Language } from '../../../../../modules/tests/communication-preferences/communication-preferences.model';
+import { Language } from '../../../../modules/tests/communication-preferences/communication-preferences.model';
 
 @Component({
   selector: 'privacy-notice',
