+++ src/pages/communication-page/components/provided-email/__tests__/provided-email.spec.ts
@@ -5,7 +5,7 @@ import { IonicModule } from 'ionic-angular';
 import { By } from '@angular/platform-browser';
 import { TranslateService, TranslateModule, TranslateLoader } from 'ng2-translate';
 import * as welshTranslations from '../../../../../../assets/i18n/cy.json';
-import { createTranslateLoader } from '../../../../../../app/app.module';
+import { createTranslateLoader } from '../../../../../app/app.module';
 import { Http } from '@angular/http';
 
 describe('ProvidedEmailComponent', () => {
