+++ src/pages/communication-page/components/provided-email/provided-email.ts
